View: Remove !! assertion of often null nullable
This makes it so the getter FragmentLandingPage::binding no longer
asserts that '_binding' is non-null - for some phones, this would crash
the program.

Though '_binding' logically should never actually be null when
FragmentLandingPage::binding might be called, it for some reason is. I
would assume this probably is caused by the loading and unloading of
fragments combined with multithreading. Sometimes, the code that will
use 'binding' will be run before the code which sets '_binding'.

Whatever the cause, this prevents the crashes - seemingly without
altering the app in any other way, from an end-user perspective at
least.