View: Add image acquisition buttons on suggestion fragment
This makes it a lot more convenient to take an image, requiring only a
single click to start the process. I had observed users spending a
considerate amount of time in stasis after making a move, seemingly
attempting to compute how to perform their next action. Having the
option to acquire an image presented right in front of them hopefully
helps with this.